api/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.7.1=py36_0
astor=0.8.0=py36_0
atomicwrites=1.3.0=py_0
attrs=19.1.0=py36_1
backcall=0.1.0=py36_0
blas=1.0=mkl
bleach=3.1.0=py36_0
ca-certificates=2019.6.16=hecc5488_0
certifi=2019.6.16=py36_1
cffi=1.12.3=py36h7a1dbc1_0
click=7.0=py36_0
colorama=0.4.1=py36_0
cuda90=1.0=0
cudatoolkit=10.0.130=0
cudnn=7.6.0=cuda10.0_0
cycler=0.10.0=py36h009560c_0
decorator=4.4.0=py36_1
defusedxml=0.6.0=py_0
entrypoints=0.3=py36_0
flake8=3.7.8=py36_1
flake8-polyfill=1.0.2=py_0
flake8-quotes=2.1.0=py_0
flask=1.0.2=py36_1
freetype=2.9.1=ha9979f8_1
gast=0.2.2=py36_0
grpcio=1.16.1=py36h351948d_1
h5py=2.9.0=py36h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
imageio=2.5.0=py36_0
importlib_metadata=0.18=py36_0
intel-openmp=2019.4=245
ipykernel=5.1.2=py36h39e3cac_0
ipython=7.8.0=py36h39e3cac_0
ipython_genutils=0.2.0=py36h3c5d0ee_0
itsdangerous=1.1.0=py36_0
jedi=0.15.1=py36_0
jinja2=2.10.1=py36_0
jpeg=9b=hb83a4c4_2
jsonschema=3.0.2=py36_0
jupyter_client=5.3.1=py_0
jupyter_core=4.5.0=py_0
keras=2.2.4=0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py36_0
keras-gpu=2.2.4=0
keras-preprocessing=1.0.5=py36_0
kiwisolver=1.1.0=py36ha925a31_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.8.0=h7bd577a_0
libsodium=1.0.16=h9d3ae62_0
libtiff=4.0.10=hb898794_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.1.1=py36_0
markupsafe=1.1.1=py36he774522_0
matplotlib=3.0.2=py36hc8f65d3_0
mccabe=0.6.1=py_1
mistune=0.8.4=py36he774522_0
mkl=2019.4=245
mkl-service=2.0.2=py36he774522_0
mkl_fft=1.0.14=py36h14836fe_0
mkl_random=1.0.2=py36h343c172_0
more-itertools=7.2.0=py_0
msys2-conda-epoch=20160418=1
nbconvert=5.5.0=py_0
nbformat=4.4.0=py36h3a5bc1b_0
ninja=1.9.0=py36h74a9793_0
notebook=5.7.4=py36_0
numpy=1.16.4=py36h19fb1c0_0
numpy-base=1.16.4=py36hc3f5095_0
olefile=0.46=py36_0
opencv-contrib-python=4.1.0.25=pypi_0
opencv-python=4.1.0.25=pypi_0
openssl=1.1.1c=hfa6e2cd_0
packaging=19.0=py_0
pandas=0.23.4=py36h830ac7b_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
parso=0.5.1=py_0
pep8-naming=0.8.2=py_0
pickleshare=0.7.5=py36_0
pillow=6.1.0=py36hdc69c19_0
pip=19.2.2=py36_0
pluggy=0.12.0=py_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.9=py36_0
protobuf=3.8.0=py36h33f27b4_0
py=1.8.0=py_0
pycodestyle=2.5.0=py_0
pycparser=2.19=py36_0
pyflakes=2.1.1=py_0
pygments=2.4.2=py_0
pyparsing=2.4.2=py_0
pyqt=5.9.2=py36h6538335_2
pyreadline=2.1=py36_1
pyrsistent=0.14.11=py36he774522_0
pytesseract=0.3.0=py_0
pytest=5.1.2=py36_0
python=3.6.8=h9f7ef89_0
python-dateutil=2.8.0=py36_0
pytorch=1.0.1=py3.6_cuda100_cudnn7_1
pytz=2019.2=py_0
pywinpty=0.5.5=py36_1000
pyyaml=5.1.2=py36he774522_0
pyzmq=18.1.0=py36ha925a31_0
qt=5.9.7=vc14h73c81de_0
scipy=1.3.1=py36h29ff71c_0
send2trash=1.5.0=py36_0
setuptools=41.0.1=py36_0
sip=4.19.8=py36h6538335_0
six=1.12.0=py36_0
sqlite=3.29.0=he774522_0
tensorboard=1.14.0=py36he3c9ec2_0
tensorflow=1.14.0=gpu_py36h305fd99_0
tensorflow-base=1.14.0=gpu_py36h55fc52a_0
tensorflow-estimator=1.14.0=py_0
tensorflow-gpu=1.14.0=h0d30ee6_0
termcolor=1.1.0=py36_1
terminado=0.8.2=py36_0
testpath=0.4.2=py36_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.2.2=py_3
tornado=5.1.1=py36hfa6e2cd_0
traitlets=4.3.2=py36h096827d_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.15.26706=h3a45250_4
wcwidth=0.1.7=py36h3d5aa90_0
webencodings=0.5.1=py36_1
werkzeug=0.15.5=py_0
wheel=0.33.4=py36_0
wincertstore=0.2=py36h7fe50ca_0
winpty=0.4.3=4
wrapt=1.11.2=py36he774522_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
zeromq=4.3.1=h33f27b4_3
zipp=0.6.0=py_0
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
